NDArray M6: Int/Boolean/Float element types with ops, reductions, boolean indexing, and where (#79)
* Initial plan

* Implement NDArray M6: Int/Boolean/Float ops, reductions, boolean indexing, where

Agent-Logs-Url: https://github.com/Quafadas/vecxt/sessions/039e1616-72d6-4f72-b5a0-13322445a4cc


* .

* .

* Fix native compilation: BooleanArrays scope conflict and @targetName erasure conflicts

Issue 1: Remove import vecxt.BooleanArrays.* from ndarrayBooleanOps.scala to
avoid scope conflict where local NDArray[Boolean] extension methods (any, all,
Replace all delegations with inline while loops.
Also fix in-place ops to check isColMajor instead of isContiguous so that
row-major transposed arrays correctly throw InvalidNDArray.

Issue 2: Add @targetName annotations to all extension methods on NDArray[Int]
and NDArray[Float] that produce type-erasure 'double definition' conflicts on
Scala Native when exported together (NDArray[T] erases to NDArray for all T).
- ndarrayIntOps.scala: annotate binary ops, unary ops, in-place ops, comparisons
- ndarrayIntReductions.scala: add import + annotate all full/axis reductions
- ndarrayFloatOps.scala: annotate binary ops, unary ops, in-place ops, comparisons
- ndarrayFloatReductions.scala: add import + annotate all full/axis reductions



* massive refactor

* fixy fixy

* fmt

* fmt fix

* .

* .

---------

+++ vecxt/src/ndarrayReductions.scala
@@ -2,9 +2,9 @@ package vecxt
 
 import vecxt.BoundsCheck.BoundsCheck
 import vecxt.broadcast.ShapeMismatchException
+import vecxt.doublearrays.*
 import vecxt.matrix.*
 import vecxt.ndarray.*
-import vecxt.arrays.*
 
 /** Private helper object for delegating to Array[Double] and Matrix[Double] extension methods.
   *
@@ -13,7 +13,7 @@ import vecxt.arrays.*
   * `Array[Double]`.
   */
 private object NDArrayReductionHelpers:
-  import vecxt.arrays.*
+  import vecxt.doublearrays.*
   import vecxt.DoubleMatrix.*
 
   inline def mean(d: Array[Double]): Double = d.mean
@@ -118,19 +118,19 @@ object NDArrayReductions:
     * `initial` should be `Double.NegativeInfinity` for argmax, `Double.PositiveInfinity` for argmin. `compare(newVal,
     * bestSoFar)` should return `true` when `newVal` is a better candidate.
     *
-    * Output is `NDArray[Double]` of integral indices (Double for type consistency).
+    * Output is `NDArray[Int]` of integral indices.
     */
   private[NDArrayReductions] inline def argReduceAxis(
       a: NDArray[Double],
       axis: Int,
       inline initial: Double,
       inline compare: (Double, Double) => Boolean
-  ): NDArray[Double] =
+  ): NDArray[Int] =
     val outShape = removeAxis(a.shape, axis)
     val outStrides = colMajorStrides(outShape)
     val outN = shapeProduct(outShape)
     val bestVals = Array.fill(outN)(initial)
-    val outIdx = new Array[Double](outN)
+    val outIdx = new Array[Int](outN)
 
     val n = a.numel
     val ndim = a.ndim
@@ -156,7 +156,7 @@ object NDArrayReductions:
       val v = a.data(posIn)
       if compare(v, bestVals(posOut)) then
         bestVals(posOut) = v
-        outIdx(posOut) = axisCoord.toDouble
+        outIdx(posOut) = axisCoord
       end if
       j += 1
     end while
@@ -322,15 +322,15 @@ object NDArrayReductions:
       reduceAxis(a, axis, 1.0, _ * _)
     end product
 
-    /** Argmax along axis `axis`. Returns NDArray[Double] of indices (values are integral). */
-    inline def argmax(axis: Int): NDArray[Double] =
+    /** Argmax along axis `axis`. Returns NDArray[Int] of indices. */
+    inline def argmax(axis: Int): NDArray[Int] =
       if axis < 0 || axis >= a.ndim then throw InvalidNDArray(s"Axis $axis out of range [0, ${a.ndim})")
       end if
       argReduceAxis(a, axis, Double.NegativeInfinity, _ > _)
     end argmax
 
-    /** Argmin along axis `axis`. Returns NDArray[Double] of indices (values are integral). */
-    inline def argmin(axis: Int): NDArray[Double] =
+    /** Argmin along axis `axis`. Returns NDArray[Int] of indices. */
+    inline def argmin(axis: Int): NDArray[Int] =
       if axis < 0 || axis >= a.ndim then throw InvalidNDArray(s"Axis $axis out of range [0, ${a.ndim})")
       end if
       argReduceAxis(a, axis, Double.PositiveInfinity, _ < _)
